Crate remote_settings

Re-exports

pub use error::RemoteSettingsError;

pub use error::Result;

pub use client::Attachment;

pub use client::Client;

pub use client::GetItemsOptions;

pub use client::RemoteSettingsRecord;

pub use client::RemoteSettingsResponse;

pub use client::RsJsonObject;

pub use client::SortOrder;

pub use config::RemoteSettingsConfig;

pub use config::RemoteSettingsServer;

Modules

cache

client

config

This module defines the custom configurations that consumers can set. Those configurations override default values and can be used to set a custom server, collection name, and bucket name. The purpose of the configuration parameters are to allow consumers an easy debugging option, and the ability to be explicit about the server.

error

Structs

RemoteSettings