remote_settings
Module client

Structs

Attachment

Attachment metadata that can be optionally attached to a [Record]. The [location] should included in calls to Client::get_attachment.

ChangesetResponse

Client

A simple HTTP client that can retrieve Remote Settings data using the properties by [ClientConfig]. Methods defined on this will fetch data from <base_url>/buckets/<bucket_name>/collections/<collection_name>/

CollectionMetadata

CollectionSignature

GetItemsOptions

Options for requests to endpoints that return multiple items.

RemoteSettingsClient

Internal Remote settings client API

RemoteSettingsRecord

A parsed Remote Settings record. Records can contain arbitrary fields, so clients are required to further extract expected values from the [fields] member.

RemoteSettingsResponse

Data structure representing the top-level response from the Remote Settings. [last_modified] will be extracted from the etag header of the response.

ViaductApiClient

Client for Remote settings API requests

Enums

SortOrder

The order in which to return items.

Traits

ApiClient

Type Aliases

RsJsonObject