Struct remote_settings::client::GetItemsOptions

pub struct GetItemsOptions { /* private fields */ }

Options for requests to endpoints that return multiple items.

Implementations

impl GetItemsOptions

pub fn new() -> Self

Creates an empty option set.

pub fn eq( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is equal to the given value.

field can be a simple or dotted field name, like author or author.name. value can be a bare number or string (like 2 or Ben), or a stringified JSON value ("2.0", [1, 2], {"checked": true}).

pub fn not( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is not equal to the given value.

pub fn contains( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is an array that contains the given value. If value is a stringified JSON array, the field must contain all its elements.

pub fn lt( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is strictly less than the given value.

pub fn gt( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is strictly greater than the given value.

pub fn max( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is less than or equal to the given value.

pub fn min( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is greater than or equal to the given value.

pub fn like( &mut self, field: impl Into<String>, value: impl Into<String> ) -> &mut Self

Sets an option to only return items whose field is a string that contains the substring value. value can contain * wildcards.

pub fn has(&mut self, field: impl Into<String>) -> &mut Self

Sets an option to only return items that have the given field.

pub fn has_not(&mut self, field: impl Into<String>) -> &mut Self

Sets an option to only return items that do not have the given field.

pub fn sort(&mut self, field: impl Into<String>, order: SortOrder) -> &mut Self

Sets an option to return items in order for the given field.

pub fn field(&mut self, field: impl Into<String>) -> &mut Self

Sets an option to only return the given field of each item.

The special id and last_modified fields are always returned.

pub fn limit(&mut self, count: u64) -> &mut Self

Sets the option to return at most count items.

pub fn iter_query_pairs( &self ) -> impl Iterator<Item = (Cow<'_, str>, Cow<'_, str>)>

Returns an iterator of (name, value) query pairs for these options.