remote_settings

Module config

This module defines the custom configurations that consumers can set. Those configurations override default values and can be used to set a custom server, collection name, and bucket name. The purpose of the configuration parameters are to allow consumers an easy debugging option, and the ability to be explicit about the server.

Structs

BaseUrl

Url that’s guaranteed safe to use as a base

RemoteSettingsConfig

Custom configuration for the client. Currently includes the following:

RemoteSettingsConfig2

Remote settings configuration

Enums

RemoteSettingsServer

The Remote Settings server that the client should use.