remote_settings
Struct RemoteSettingsContext

pub struct RemoteSettingsContext {

    pub app_name: String,
    pub app_id: String,
    pub channel: String,
    pub app_version: Option<String>,
    pub app_build: Option<String>,
    pub architecture: Option<String>,
    pub device_manufacturer: Option<String>,
    pub device_model: Option<String>,
    pub locale: Option<String>,
    pub os: Option<String>,
    pub os_version: Option<String>,
    pub android_sdk_version: Option<String>,
    pub debug_tag: Option<String>,
    pub installation_date: Option<i64>,
    pub home_directory: Option<String>,
    pub custom_targeting_attributes: Option<Map<String, Value>>,

}

The RemoteSettingsContext object represents the parameters and characteristics of the consuming application. For remote-settings, it is used to filter locally stored records.

We always fetch all records from the remote-settings storage. Some records could have a filter_expression attached to them, which will be matched against the RemoteSettingsContext.

When set, only records where the expression is true will be returned.

Fields
app_name: String

Name of the application (e.g. “Fenix” or “Firefox iOS”)

app_id: String

Application identifier, especially for mobile (e.g. “org.mozilla.fenix”)

channel: String

The delivery channel of the application (e.g “nightly”)

app_version: Option<String>

User visible version string (e.g. “1.0.3”)

app_build: Option<String>

Build identifier generated by the CI system (e.g. “1234/A”)

architecture: Option<String>

The architecture of the device, (e.g. “arm”, “x86”)

device_manufacturer: Option<String>

The manufacturer of the device the application is running on

device_model: Option<String>

The model of the device the application is running on

locale: Option<String>

The locale of the application during initialization (e.g. “es-ES”)

os: Option<String>

The name of the operating system (e.g. “Android”, “iOS”, “Darwin”, “Windows”)

os_version: Option<String>

The user-visible version of the operating system (e.g. “1.2.3”)

android_sdk_version: Option<String>

Android specific for targeting specific sdk versions

debug_tag: Option<String>

Used for debug purposes as a way to match only developer builds, etc.

installation_date: Option<i64>

The date the application installed the app

home_directory: Option<String>

The application’s home directory

custom_targeting_attributes: Option<Map<String, Value>>

Contains attributes specific to the application, derived by the application