Crate remote_settings

Re-exports

pub use client::Attachment;

pub use client::RemoteSettingsRecord;

pub use client::RemoteSettingsResponse;

pub use client::RsJsonObject;

pub use config::RemoteSettingsConfig;

pub use config::RemoteSettingsConfig2;

pub use config::RemoteSettingsServer;

pub use error::ApiResult;

pub use error::RemoteSettingsError;

pub use error::Result;

Modules

cache

client

config

This module defines the custom configurations that consumers can set. Those configurations override default values and can be used to set a custom server, collection name, and bucket name. The purpose of the configuration parameters are to allow consumers an easy debugging option, and the ability to be explicit about the server.

error

schema

service

storage

Macros

packaged_attachments

packaged_collections

Structs

RemoteSettings

RemoteSettingsClient

Client for a single Remote Settings collection

RemoteSettingsContext

The RemoteSettingsContext object represents the parameters and characteristics of the consuming application. For remote-settings, it is used to filter locally stored records.

RemoteSettingsService

Application-level Remote Settings manager.