Crate remote_settings

Re-exports

pub use client::Attachment;

pub use client::RemoteSettingsRecord;

pub use client::RemoteSettingsResponse;

pub use client::RsJsonObject;

pub use config::BaseUrl;

pub use config::RemoteSettingsConfig;

pub use config::RemoteSettingsConfig2;

pub use config::RemoteSettingsServer;

pub use context::RemoteSettingsContext;

pub use error::ApiResult;

pub use error::RemoteSettingsError;

pub use error::Result;

Modules

cache

client

config

This module defines the custom configurations that consumers can set. Those configurations override default values and can be used to set a custom server, collection name, and bucket name. The purpose of the configuration parameters are to allow consumers an easy debugging option, and the ability to be explicit about the server.

context

error

schema

service

storage

Macros

packaged_attachments

packaged_collections

trace

Structs

RemoteSettings

RemoteSettingsClient

Client for a single Remote Settings collection

RemoteSettingsService

Application-level Remote Settings manager.